+++ configserver/src/main/java/com/yahoo/vespa/config/server/application/VersionDoesNotExistException.java
@@ -1,11 +1,13 @@
 // Copyright 2016 Yahoo Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-package com.yahoo.vespa.config.server;
+package com.yahoo.vespa.config.server.application;
 
 /**
  * @author Vegard Sjonfjell
  */
 public final class VersionDoesNotExistException extends RuntimeException {
+
     public VersionDoesNotExistException(String message) {
         super(message);
     }
+
 }
