+++ configserver/src/test/java/com/yahoo/vespa/config/server/application/ApplicationSetTest.java
@@ -9,9 +9,6 @@ import java.util.Optional;
 import com.yahoo.config.provision.ApplicationId;
 import com.yahoo.config.provision.Version;
 import com.yahoo.vespa.config.server.ModelStub;
-import com.yahoo.vespa.config.server.VersionDoesNotExistException;
-import com.yahoo.vespa.config.server.application.Application;
-import com.yahoo.vespa.config.server.application.ApplicationSet;
 import com.yahoo.vespa.config.server.monitoring.MetricUpdater;
 import org.junit.Before;
 import org.junit.Test;
